.then() after .or() also fails w/ err msg from branch...
that got furthest.

success continuations now take 2 more arguments: the furthest in the
stream any subparser of parser has ever gotten before backtracing, and
the corresponding `expected` message.

they're used by the `success` continuation that .then() passes to the
`self` parser, the arguments are then used like in .or(): to compare
against them in the failure continuation passed the `next` parser.
.or() and .then() both also have to pass them forward in the success
continuations too, of course.

so factor out from .or() HOFs that construct success and failure
continuations to pass on that will always preserve the error message
from the branch that got furthest, constructed from the onSuccess and
onFailure continuations.

also the base parsers, if they succeed, no subparser in them has ever
failed, so for `furthestFailedI` they pass in -1 (less even than 0, the
least far you could possibly fail), which any actual failure is
guaranteed to be at a larger index than and hence ignore their
`expected` message (or lack of one).