Accept varargs in Parsimmon.seq()
The common case for seq is to pass a fixed number of parsers, so using
varargs rather than an explicit array is syntactically cleaner (as
well as less surprising IMHO).  I've kept the previous
array-as-first-arg behavior for backwards compatability.
Alternatively, that could be removed and anyone wishing to pass a
variable number of args can do:

seq.apply(null, [args])