allow failures to aggregate messages at equal indices
This approach is lifted wholesale from Parsec.  It turns out that it's
relatively common for several parsers to fail at the same index, which
indicates more than one possible way the user could fix the parse error.
This changes the .expected keys from strings to arrays of strings which
aggregate equal-index parse errors.  They can still be passed to
formatExpected(), so this should be non-breaking for anyone who was
simply passing the expected result in - but it will break for anyone who
was accessing the string directly.