Update Travis CI configuration
The comment with the URL seemed to point to a test dashboard that
doesn't exist anymore, so I deleted it.

The versions of Node.js being tested on are also rather old (so I added
some newer ones).

The tests were running on the legacy Travis infrastructure (so I added
Sudo: false as the docs recommend
http://docs.travis-ci.com/user/migrating-from-legacy/ )

This might fix at least part of the constantly failing CI builds and
make them useful again.