Create methods for marking line and column offsets
PEG.js and Jison give locations not only as a character offset but also
as line and column numbers. They help make errors human-readable and are
necessary for various other tools e.g. source maps.

Calculating line and column positions of everything after parsing has
finished mixes concerns and is inefficient.

This adds that functionality to Parsimmon, but optional such that it
won't reduce performance when not used.  I stuck approximately to the
idioms established by _.mark and _.index.

Fixes #54.