fix: support single carriage return
These type of line breaks are ancient now, they were seen on Mac OS 9
and earlier version. Those files still exist and sometimes we'll have to
parse them.

The behavior before the fix:

If the file contained ^M line endings the marks will be off

After the fix:

If the file contained ^M line endings the marks should be correct